+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 7.3.11 (3/23/2020 - 
 * kafka: set linger.ms to 5ms by default to improve batching under load
+* security: add X-Frame-Options in WebSecurityInterceptor
 
 ### 7.3.10 (3/12/2020 - 3/19/2020)
 * log: increase max length of error message to 1000, to show more info if needed in notification